backend/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
ca-certificates=2022.10.11=hecd8cb5_0
certifi=2022.9.24=py38hecd8cb5_0
click=8.0.4=py38hecd8cb5_0
dataclasses=0.8=pyh6d0b6a4_7
flask=2.1.3=py38hecd8cb5_0
flask_cors=3.0.10=pyhd3eb1b0_0
importlib-metadata=4.11.3=py38hecd8cb5_0
itsdangerous=2.0.1=pyhd3eb1b0_0
jinja2=3.1.2=py38hecd8cb5_0
joblib=1.2.0=pypi_0
libcxx=14.0.6=h9765a3e_0
libffi=3.3=hb1e8313_2
markupsafe=2.1.1=py38hca72f7f_0
modal=0.4.1=pypi_0
ncurses=6.3=hca72f7f_3
nptsne=1.2.0=pypi_0
numpy=1.23.5=pypi_0
openssl=1.1.1s=hca72f7f_0
pandas=1.5.2=pypi_0
pip=22.2.2=py38hecd8cb5_0
python=3.8.15=h4319210_0
python-dateutil=2.8.2=pypi_0
python-dotenv=0.21.0=pypi_0
pytz=2022.6=pypi_0
readline=8.2=hca72f7f_0
scikit-learn=1.1.3=pypi_0
scipy=1.9.3=pypi_0
setuptools=65.5.0=py38hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0.post1=pypi_0
sqlite=3.40.0=h880c91c_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h5d9f67b_0
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.6=hca72f7f_0
zipp=3.8.0=py38hecd8cb5_0
zlib=1.2.13=h4dc903c_0
